+++ js/main.js
@@ -1,6 +1,6 @@
 import { getRandomPositiveInteger } from './utils/get-random-positive-integer.js';
 import { getRandomPositiveFloat } from './utils/get-random-positive-float.js';
-import{ generationOneCard } from './generation-card.js';
+//import{ generationOneCard } from './generation-card.js';
 import { shutDownDocument } from './no-active-document.js';
 import { turningOnDocument } from './active-document.js';
 import { checkTitleValidity } from './form-utils/check-title-validity.js';
@@ -10,6 +10,7 @@ import { setPrice } from './form-utils/set-Price.js';
 import { timeinTimeout } from './form-utils/timein-timout.js';
 import { createLoader, sendData } from './load.js';
 import { showAlert } from './show-alert.js';
+import { renderTagMarkers } from './form-utils/render-tag-markers.js';
 getRandomPositiveFloat(1.2323, 2.1122);
 getRandomPositiveInteger(1,10);
 
@@ -104,21 +105,10 @@ const marker = L.marker(
 
 const data = createLoader(showAlert);
 
-data.then((array) => array.forEach((tag) => {
-  const iconTag = L.icon({
-    iconUrl: 'img/pin.svg',
-    iconSize: [52, 52],
-    iconAnchor: [26, 52],
-  });
-  const tagMarker = L.marker(
-    tag.location,
-    {
-      draggable: true,
-      icon: iconTag,
-    },
-  );
-  tagMarker.addTo(map).bindPopup(generationOneCard(tag));
-}));
+data.then((array) => {
+  console.log(array);
+  renderTagMarkers(array, map);
+});
 
 marker.addTo(map);
